chore: move FakeCoordinator to tailnettest
